Add files via upload
I had the opportunity to collaborate with my friends Krishil Rana, Aditya Pachchigar, and Bhargav Patel on exciting challenges and the development of innovative projects.

The problem statement was to create a GUI application for a chess board that would place pieces based on given data and find the closest safe position for the white piece. And we were supposed to optimize the data structure as well as the algorithm used for it.

One of the aspects of the hackathon that I appreciated the most was the opportunity to learn from industry experts who were on hand to provide guidance and support. I also valued the opportunity to interact with other people who shared my interests.

Everyone worked extremely hard and went above and beyond to create something truly exceptional, creating an electrifying atmosphere.
The level of ingenuity and creativity on display was astounding, and I left the event feeling energized and determined to continue pursuing my passion for technology.

Fortunately, my team and I won First Place in the Track "Smart Navigator" and received a Certificate of Achievement, a coffee mug, and a Rs. 20,000 Amazon Gift Card.